城市三字代码、客户需求编辑select2

+++ we-web-parent/web-linyun-airline/src/main/java/com/linyun/airline/admin/order/international/service/InternationalService.java
@@ -718,7 +718,9 @@ public class InternationalService extends BaseService<TUpOrderEntity> {
 			List<TVisitorInfoEntity> query = dbDao
 					.query(TVisitorInfoEntity.class, Cnd.where("pnrid", "=", pnrid), null);
 			//删除原有的
-			dbDao.delete(query);
+			if (!Util.isEmpty(query)) {
+				dbDao.delete(query);
+			}
 			List<TVisitorInfoEntity> visitors = new ArrayList<TVisitorInfoEntity>();
 			for (int i = 1; i <= map.size(); i++) {
 				String[] row = map.get(i);
